requirements.txt
torch
torchvision
scikit-learn
matplotlib
Pillow
tqdm
mlflow
fastapi
uvicorn
onnx
onnxruntime
pandas
seaborn
numpy<1.24
protobuf>=3.20,<3.21
python-multipart